requirements.txt
# Django==3.2.25
# django-authtools==1.6.0
# django-braces==1.13.0
# django-cors-headers==4.1.0
# django-crispy-forms==1.7.2
# django-debug-toolbar==1.10.1
# django-environ==0.4.5
# django-spa==0.3.6
# djangorestframework==3.15.1
# python-decouple>=3.8  # For environment variables management
# # Database (if using PostgreSQL in production)
# psycopg2-binary==2.9.9

# # Security & Performance
# gunicorn==21.2.0  # WSGI server for production
# whitenoise==6.3.0  # Static file serving

# # Testing
# pytest==7.4.3
# pytest-django==4.5.0
# dj-database-url==3.0.1

# # Django 5.1 LTS (Latest Long Term Support)
# Django==5.1.3

# # Django REST Framework
# djangorestframework==3.15.2

# # CORS Headers
# django-cors-headers==4.6.0

# # Database URL utility
# dj-database-url==2.2.0

# # Environment Variables
# python-decouple==3.8

# # PostgreSQL adapter
# psycopg2-binary==2.9.10

# # Production server
# gunicorn==23.0.0

# uvicorn==0.38.0

# # Static files
# whitenoise==6.8.2


# Core Django
Django==4.2.17

# Django REST Framework
djangorestframework==3.14.0

# CORS Headers
django-cors-headers==4.3.1

# Database URL utility
dj-database-url==2.1.0

# Environment Variables
python-decouple==3.8

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

uvicorn==0.38.0

pandas==2.3.3

sqlalchemy==1.4 